Moved Facility Class to new folder and imported it.
