+++ Facility.py
@@ -1,0 +1,20 @@
+class Facility:
+    def __init__(self, facilityName, adultMaleCount, adultFemaleCount, juvenileMaleCount, juvenileFemaleCount,
+                 inCountOutCountMale, inCountOutCountFemale, workersMale, workersFemale, furloughMale,
+                 furloughFemale, openWardMale, openWardFemale, emergTripsMale, emergTripsFemale):
+
+        self.facilityName = facilityName
+        self.adultMaleCount = adultMaleCount
+        self.adultFemaleCount = adultFemaleCount
+        self.juvenileMaleCount = juvenileMaleCount
+        self.juvenileFemaleCount = juvenileFemaleCount
+        self.inCountOutCountMale = inCountOutCountMale
+        self.inCountOutCountFemale = inCountOutCountFemale
+        self.workersMale = workersMale
+        self.workersFemale = workersFemale
+        self.furloughMale = furloughMale
+        self.furloughFemale = furloughFemale
+        self.openWardMale = openWardMale
+        self.openWardFemale = openWardFemale
+        self.emergTripsMale = emergTripsMale
+        self.emergTripsFemale = emergTripsFemale
